Implement SSSP problem using Dijkstra's algorithm and relocated it under P folder
