+++ Problems/NPComplete/NPC_PARTITION/Visualizations/PartitionDefaultVisualization.cs
@@ -3,7 +3,6 @@ using API.Interfaces.Graphs.GraphParser;
 using API.Interfaces.JSON_Objects;
 using API.Interfaces.JSON_Objects.Graphs;
 using SPADE;
-using API.Problems.NPComplete.NPC_SHORTESTPATH.Solvers;
 using API.Problems.NPComplete.NPC_PARTITION.Solvers;
 
 namespace API.Problems.NPComplete.NPC_PARTITION.Visualizations;
