+++ Problems/P/P_SSSP/SSSP_Class.cs
@@ -1,23 +1,23 @@
 ﻿using API.Interfaces;
-using API.Problems.NPComplete.NPC_SHORTESTPATH.Solvers;
-using API.Problems.NPComplete.NPC_SHORTESTPATH.Verifiers;
-using API.Problems.NPComplete.NPC_SHORTESTPATH.Visualizations;
+using API.Problems.P.P_SSSP.Solvers;
+using API.Problems.P.P_SSSP.Verifiers;
+using API.Problems.P.P_SSSP.Visualizations;
 using SPADE;
 using System.Linq;
 using System.Collections.Generic;
 using API.Interfaces.Graphs;
 using System.Text;
 
-namespace API.Problems.NPComplete.NPC_SHORTESTPATH;
+namespace API.Problems.P.P_SSSP;
 
-class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, ShortestPathVisualization, UtilCollectionGraph>
+class SSSP: IGraphProblem<SSSPSolver, SSSPVerifier, SSSPVisualization, UtilCollectionGraph>
 {
 
     // --- Fields ---
-    public string problemName { get; } = "Shortest Path Problem";
+    public string problemName { get; } = "Single Source Shortest Path Problem";
     public string problemLink { get; } = "https://en.wikipedia.org/wiki/Shortest_path_problem";
-    public string formalDefinition { get; } = "For a weighted graph G= (V,E) with weight function w:E->\\(\\mathbb{R}\\), find the shortest path source s to target t, where path length is the sum of edge weights"; //TODO figure out what it's supposed to be
-    public string problemDefinition { get; } = "Shortest Path Problem is the problem of determining the shortest path from a given source vertex to all other vertices such that the sum of the weights of its inherent edges is minimized.";
+    public string formalDefinition { get; } = "For a weighted graph G= (V,E) with non-negative edge weights and source vertex s \u2208 V, find the shortest path distance from s to every other vertex v \u2208 V, where path length is defined as the sum of edge weights along the path."; 
+    public string problemDefinition { get; } = "Single Source Shortest Path (SSSP) in a weighted graph is the problem of finding the shortest path from a given source vertex s in the graph, such that the sum of edge weights along each path is minimized.";
     public string source { get; } = "N/A";
     public string sourceLink { get; } = "N/A";
     private static string _defaultInstance =
@@ -32,11 +32,11 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
     public bool isWeighted { get; private set; }
     private List<string> _nodes = new List<string>();
     private List<KeyValuePair<string, string>> _edges = new List<KeyValuePair<string, string>>();
-    public DijkstraSolver defaultSolver { get; } = new DijkstraSolver();
-    public ShortestPathVerifier defaultVerifier { get; } = new ShortestPathVerifier();
-    public ShortestPathVisualization defaultVisualization { get; } = new ShortestPathVisualization();
+    public SSSPSolver defaultSolver { get; } = new SSSPSolver();
+    public SSSPVerifier defaultVerifier { get; } = new SSSPVerifier();
+    public SSSPVisualization defaultVisualization { get; } = new SSSPVisualization();
     public UtilCollectionGraph graph { get; set; }
-    public string[] contributors { get; } = { "Tiger Sant", "Malaya Witt", "Rajit Nilkar", "Scott Barfuss" };
+    public string[] contributors { get; } = { "Rajit Nilkar", "Scott Barfuss" };
     
     // --- Properties ---
     public List<string> nodes
@@ -51,9 +51,9 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
     }
 
     // --- Methods Including Constructors ---
-    public SHORTESTPATH() : this(_defaultInstance) { }
+    public SSSP() : this(_defaultInstance) { }
 
-    public SHORTESTPATH(string GInput)
+    public SSSP(string GInput)
     {
         instance = GInput;
 
@@ -87,7 +87,7 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
             explicitTarget = outerTerms[2];
         }
 
-            GraphParseResult graphParse = ParseGraph(graphInput);
+        GraphParseResult graphParse = ParseGraph(graphInput);
         UtilCollection nodeCollection = graphParse.Parser["N"] ?? throw new InvalidOperationException("Failed to parse N (nodes).");
         UtilCollection edgeCollection = graphParse.Parser["E"] ?? throw new InvalidOperationException("Failed to parse E (edges).");
 
@@ -135,7 +135,7 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
             }
         }
 
-        throw new InvalidOperationException("Failed to parse SHORTESTPATH instance.", lastError);
+        throw new InvalidOperationException("Failed to parse SSSP instance.", lastError);
     }
 
     private static void ValidateInstance(
@@ -163,7 +163,7 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
                 throw new InvalidOperationException($"Edge target '{edge.To}' is not in N.");
 
             if (edge.Weight < 0)
-                throw new InvalidOperationException("SHORTESTPATH does not allow negative edge weights.");
+                throw new InvalidOperationException("SSSP using Dijkstra's algorithm does not allow negative edge weights.");
         }
     }
 
@@ -188,7 +188,7 @@ class SHORTESTPATH: IGraphProblem<DijkstraSolver, ShortestPathVerifier, Shortest
             int weight = int.Parse(rawEdge[1].ToString());
 
             if (weight < 0)
-                throw new InvalidOperationException($"SHORTESTPATH does not allow negative edge weights. Found edge weight: {weight}");
+                throw new InvalidOperationException($"SSSP using Dijkstra's algorithm does not allow negative edge weights. Found edge weight: {weight}");
 
             return new ParsedEdge(GetFrom(endpoints), GetTo(endpoints), weight);
         }
