+++ Problems/P/P_SSSP/Visualizations/SSSPVisualization.cs
@@ -2,22 +2,22 @@ using API.Interfaces;
 using API.Interfaces.Graphs.GraphParser;
 using API.Interfaces.JSON_Objects;
 using API.Interfaces.JSON_Objects.Graphs;
-using API.Problems.NPComplete.NPC_SHORTESTPATH.Solvers;
+using API.Problems.P.P_SSSP.Solvers;
 
-namespace API.Problems.NPComplete.NPC_SHORTESTPATH.Visualizations;
+namespace API.Problems.P.P_SSSP.Visualizations;
 
-class ShortestPathVisualization : IVisualization<SHORTESTPATH>
+class SSSPVisualization : IVisualization<SSSP>
 {
-    public string visualizationName { get; } = "Shortest Path Visualization";
-    public string visualizationDefinition { get; } = "Visualizes the BFS and Dijkstra's algorithm";
+    public string visualizationName { get; } = "Single Source Shortest Path Visualization";
+    public string visualizationDefinition { get; } = "Visualizes the Single Source Shortest Path problem for non-negative weighted directed cyclic graphs using Dijkstra's algorithm";
     public string source { get; } = "";
     public string[] contributors { get; } = { "Rajit Nilkar", "Scott Barfuss", "Tiger Sant", "Malaya Witt"};
     public string visualizationType => "Graph D3";
-    public ISolver solver { get; } = new DijkstraSolver();
+    public ISolver solver { get; } = new SSSPSolver();
 
-    public ShortestPathVisualization() { }
+    public SSSPVisualization() { }
 
-    public API_JSON visualize(SHORTESTPATH problem)
+    public API_JSON visualize(SSSP problem)
     {
         // For simplicity, we will just return a JSON representation of the graph
         // In a real implementation, this would be more complex and would include visual elements
@@ -26,7 +26,7 @@ class ShortestPathVisualization : IVisualization<SHORTESTPATH>
 
     // SolvedVisualization: takes a problem instance and a solution certificate,
     // and returns a visualization of the problem instance with the solution highlighted
-    public API_JSON SolvedVisualization(SHORTESTPATH problem, string solution)
+    public API_JSON SolvedVisualization(SSSP problem, string solution)
     {
         if (string.IsNullOrWhiteSpace(solution) || solution.Trim() == "{}")
             // No path found, return graph with no highlights
@@ -69,56 +69,56 @@ class ShortestPathVisualization : IVisualization<SHORTESTPATH>
         return graph;
     }
 
-    public API_JSON AlternativePathsVisualization(SHORTESTPATH problem, string solution)
-    {
-        if (string.IsNullOrWhiteSpace(solution) || solution.Trim() == "{}")
-            // No path found, return graph with no highlights
-            return visualize(problem);
+    //public API_JSON AlternativePathsVisualization(SSSP problem, string solution)
+    //{
+    //    if (string.IsNullOrWhiteSpace(solution) || solution.Trim() == "{}")
+    //        // No path found, return graph with no highlights
+    //        return visualize(problem);
         
-        List<string> path;
-        try
-        {
-            // Parse the solution as a path
-            path = GraphParser.parseNodeListWithStringFunctions(solution);
-        }
-        catch
-        {
-            // Invalid solution format, return graph with no highlights
-            return visualize(problem);
-        }
+    //    List<string> path;
+    //    try
+    //    {
+    //        // Parse the solution as a path
+    //        path = GraphParser.parseNodeListWithStringFunctions(solution);
+    //    }
+    //    catch
+    //    {
+    //        // Invalid solution format, return graph with no highlights
+    //        return visualize(problem);
+    //    }
 
-        API_GraphJSON graph = problem.graph.ToAPIGraph(); // Convert to API graph format
+    //    API_GraphJSON graph = problem.graph.ToAPIGraph(); // Convert to API graph format
 
-        var pathNodes = new HashSet<string>(path);
-        for (int i = 0; i < graph.nodes.Count; i++)
-            if (pathNodes.Contains(graph.nodes[i].name))
-                graph.nodes[i].color = "SolutionAlt";
+    //    var pathNodes = new HashSet<string>(path);
+    //    for (int i = 0; i < graph.nodes.Count; i++)
+    //        if (pathNodes.Contains(graph.nodes[i].name))
+    //            graph.nodes[i].color = "SolutionAlt";
 
         
-        var pathEdges = new HashSet<(string u, string v)>();
-        for (int i = 0; i < path.Count - 1; i++)
-            pathEdges.Add((path[i], path[i + 1]));
+    //    var pathEdges = new HashSet<(string u, string v)>();
+    //    for (int i = 0; i < path.Count - 1; i++)
+    //        pathEdges.Add((path[i], path[i + 1]));
 
-        for (int i = 0; i < graph.links.Count; i++)
-        {
-            var link = graph.links[i];
-            bool isForwardPathEdge = pathEdges.Contains((link.source, link.target));
-            bool isReversePathEdge = !problem.isDirected && pathEdges.Contains((link.target, link.source));
+    //    for (int i = 0; i < graph.links.Count; i++)
+    //    {
+    //        var link = graph.links[i];
+    //        bool isForwardPathEdge = pathEdges.Contains((link.source, link.target));
+    //        bool isReversePathEdge = !problem.isDirected && pathEdges.Contains((link.target, link.source));
 
-            if (isForwardPathEdge || isReversePathEdge)
-                link.color = "SolutionAlt";
-        }
-        return graph;
-    }
+    //        if (isForwardPathEdge || isReversePathEdge)
+    //            link.color = "SolutionAlt";
+    //    }
+    //    return graph;
+    //}
 
 
-    public List<API_JSON> stepsVisualization(SHORTESTPATH problem, List<string> steps)
+    public List<API_JSON> StepsVisualization(SSSP problem, List<Object> steps)
     {
         var result = new List<API_JSON>();
 
         for (int s = 0; s < steps.Count; s++)
         {
-            string step = steps[s];
+            string step = steps[s].ToString();
 
             if(string.IsNullOrWhiteSpace(step) || step.Trim() == "{}")
             {
@@ -167,21 +167,11 @@ class ShortestPathVisualization : IVisualization<SHORTESTPATH>
             {
                 var link = graph.links[i];
                 bool isForwardPathEdge = pathEdges.Contains((link.source, link.target));
-                bool isReversePathEdge = !problem.isDirected && pathEdges.Contains((link.target, link.source));
+                bool isReversedPathEdge = !problem.isDirected && pathEdges.Contains((link.target, link.source));
 
-                if (isForwardPathEdge || isReversePathEdge)
-                    link.color = "Solution";
-                else
-                    link.color = "Background";
+                link.color = (isForwardPathEdge || isReversedPathEdge) ? "Solution" : "Background";
             }
-
             result.Add(graph);
-
-            if (s > 0 && s < steps.Count - 1)
-            {
-                // Show the previously explored path as an alternate frame between interior steps.
-                result.Add(AlternativePathsVisualization(problem, steps[s - 1]));
-            }
         }
 
         return result;
